cmd/untrash.go
package cmd

import (
	"github.com/harryzcy/mailbox-cli/internal/command"
	"github.com/spf13/cobra"
)

var commandUntrash = command.Untrash

// untrashCmd represents the untrash command
var untrashCmd = &cobra.Command{
	Use:   "untrash messageID",
	Short: "Untrash an email",
	Args:  cobra.ExactArgs(1),
	Run: func(cmd *cobra.Command, args []string) {
		messageID := args[0]

		verbose, _ := cmd.Flags().GetBool("verbose")
		result, err := commandUntrash(command.UntrashOptions{
			APIID:    cmd.Flag("api-id").Value.String(),
			Region:   cmd.Flag("region").Value.String(),
			Endpoint: cmd.Flag("endpoint").Value.String(),
			Verbose:  verbose,

			MessageID: messageID,
		})
		if err != nil {
			cmd.PrintErrln(err)
			osExit(1)
			return
		}

		cmd.Println(result)
	},
}

func init() {
	rootCmd.AddCommand(untrashCmd)
}
